added terraform sample
